# bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "ZeraAPI, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/api/Auth/Register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Register"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Auth/Login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Login"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/Cart/createCart": {
      "post": {
        "tags": [
          "Cart"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Cart"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Cart/update": {
      "put": {
        "tags": [
          "Cart"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Cart"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Cart/Delete": {
      "delete": {
        "tags": [
          "Cart"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CartIdRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/CartItem/{cartId}": {
      "get": {
        "tags": [
          "CartItem"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "cartId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/CartItem/create": {
      "post": {
        "tags": [
          "CartItem"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CartItem"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/CartItem/add": {
      "post": {
        "tags": [
          "CartItem"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/AddToCartRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/CartItem/{id}": {
      "put": {
        "tags": [
          "CartItem"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CartItem"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "CartItem"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "definitions": {
    "AddToCartRequest": {
      "type": "object",
      "properties": {
        "productId": {
          "format": "int32",
          "type": "integer"
        }
      },
      "additionalProperties": false
    },
    "Cart": {
      "type": "object",
      "properties": {
        "cartId": {
          "type": "string"
        },
        "id": {
          "type": "string"
        },
        "cartItems": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/CartItem"
          }
        }
      },
      "additionalProperties": false
    },
    "CartIdRequest": {
      "type": "object",
      "properties": {
        "cartId": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "CartItem": {
      "type": "object",
      "properties": {
        "cartItemId": {
          "format": "int32",
          "type": "integer"
        },
        "quantity": {
          "format": "int32",
          "type": "integer"
        },
        "price": {
          "format": "double",
          "type": "number"
        },
        "cartId": {
          "type": "string"
        },
        "cart": {
          "$ref": "#/definitions/Cart"
        },
        "productId": {
          "format": "int32",
          "type": "integer"
        },
        "productName": {
          "type": "string"
        },
        "product": {
          "$ref": "#/definitions/Product"
        }
      },
      "additionalProperties": false
    },
    "Login": {
      "required": [
        "password",
        "userName"
      ],
      "type": "object",
      "properties": {
        "loginId": {
          "format": "int32",
          "type": "integer"
        },
        "userName": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "Order": {
      "required": [
        "orderDate",
        "shippingAddress",
        "shippingDate",
        "totalAmount"
      ],
      "type": "object",
      "properties": {
        "orderId": {
          "format": "int32",
          "type": "integer"
        },
        "orderDate": {
          "format": "date-time",
          "type": "string"
        },
        "shippingDate": {
          "format": "date-time",
          "type": "string"
        },
        "shippingAddress": {
          "type": "string"
        },
        "totalAmount": {
          "format": "double",
          "type": "number"
        },
        "status": {
          "$ref": "#/definitions/OrderStatus"
        },
        "paymentMethodId": {
          "format": "int32",
          "type": "integer"
        },
        "paymentMethod": {
          "$ref": "#/definitions/PaymentMethod"
        },
        "id": {
          "type": "string"
        },
        "userCustomised": {
          "$ref": "#/definitions/UserCustomised"
        }
      },
      "additionalProperties": false
    },
    "OrderItem": {
      "type": "object",
      "properties": {
        "orderItemId": {
          "format": "int32",
          "type": "integer"
        },
        "quantity": {
          "format": "int32",
          "type": "integer"
        },
        "price": {
          "format": "double",
          "type": "number"
        },
        "orderId": {
          "format": "int32",
          "type": "integer"
        },
        "order": {
          "$ref": "#/definitions/Order"
        }
      },
      "additionalProperties": false
    },
    "OrderStatus": {
      "format": "int32",
      "enum": [
        0,
        1,
        2,
        3
      ],
      "type": "integer"
    },
    "PaymentMethod": {
      "required": [
        "cardNumber",
        "cvv",
        "expiryDate",
        "method",
        "nameOnCard",
        "payPalEmail"
      ],
      "type": "object",
      "properties": {
        "paymentMethodId": {
          "format": "int32",
          "type": "integer"
        },
        "method": {
          "type": "string"
        },
        "cardNumber": {
          "type": "string"
        },
        "expiryDate": {
          "type": "string"
        },
        "cvv": {
          "type": "string"
        },
        "nameOnCard": {
          "type": "string"
        },
        "payPalEmail": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "Product": {
      "required": [
        "description",
        "imageURL",
        "price",
        "productName",
        "producttype"
      ],
      "type": "object",
      "properties": {
        "productId": {
          "format": "int32",
          "type": "integer"
        },
        "productName": {
          "type": "string"
        },
        "producttype": {
          "type": "string"
        },
        "price": {
          "format": "double",
          "type": "number"
        },
        "description": {
          "type": "string"
        },
        "imageURL": {
          "type": "string"
        },
        "cartItem": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/CartItem"
          }
        },
        "orderItem": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/OrderItem"
          }
        },
        "wishListItem": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/WishListItem"
          }
        }
      },
      "additionalProperties": false
    },
    "Register": {
      "required": [
        "address",
        "email",
        "name",
        "password",
        "passwordConfirmation"
      ],
      "type": "object",
      "properties": {
        "registerId": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "passwordConfirmation": {
          "type": "string"
        },
        "address": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "UserCustomised": {
      "required": [
        "name"
      ],
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "userName": {
          "type": "string"
        },
        "normalizedUserName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "normalizedEmail": {
          "type": "string"
        },
        "emailConfirmed": {
          "type": "boolean"
        },
        "passwordHash": {
          "type": "string"
        },
        "securityStamp": {
          "type": "string"
        },
        "concurrencyStamp": {
          "type": "string"
        },
        "phoneNumber": {
          "type": "string"
        },
        "phoneNumberConfirmed": {
          "type": "boolean"
        },
        "twoFactorEnabled": {
          "type": "boolean"
        },
        "lockoutEnd": {
          "format": "date-time",
          "type": "string"
        },
        "lockoutEnabled": {
          "type": "boolean"
        },
        "accessFailedCount": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "maxLength": 100,
          "minLength": 0,
          "type": "string"
        },
        "address": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "WhishList": {
      "type": "object",
      "properties": {
        "whishListId": {
          "format": "int32",
          "type": "integer"
        },
        "id": {
          "type": "string"
        },
        "userCustomised": {
          "$ref": "#/definitions/UserCustomised"
        },
        "createdAt": {
          "format": "date-time",
          "type": "string"
        },
        "updatedAt": {
          "format": "date-time",
          "type": "string"
        },
        "wishListItems": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/WishListItem"
          }
        }
      },
      "additionalProperties": false
    },
    "WishListItem": {
      "type": "object",
      "properties": {
        "wishListItemId": {
          "format": "int32",
          "type": "integer"
        },
        "wishList": {
          "$ref": "#/definitions/WhishList"
        },
        "whishListId": {
          "format": "int32",
          "type": "integer"
        },
        "productId": {
          "format": "int32",
          "type": "integer"
        },
        "product": {
          "$ref": "#/definitions/Product"
        }
      },
      "additionalProperties": false
    }
  }
}